Move useSafeAsync() to 'hooks' directory
